+++ lib/team.mli
@@ -31,9 +31,14 @@ val members : t -> Member.t list
 type lint_report
 
 val aggregate :
-  ?okr_db:Masterdb.t -> string -> year:int -> week:int -> t list -> Report.t
-(** [aggregate admin_dir week teams] aggregates the reports of the teams for the
-    given week. *)
+  ?okr_db:Masterdb.t ->
+  string ->
+  year:int ->
+  weeks:int list ->
+  t list ->
+  Report.t
+(** [aggregate admin_dir weeks teams] aggregates the reports of the teams for the
+    given weeks. *)
 
 val lint : string -> year:int -> weeks:int list -> t list -> lint_report
 (** [lint admin_dir weeks teams] generates a [lint_report] for the teams at the